+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/providers/providers.container.tsx
@@ -15,7 +15,7 @@
 import * as React from 'react'
 import { hot } from 'react-hot-loader'
 
-import ThemeContainer from '../../react-component/container/theme-container'
+import ThemeContainer from '../../react-component/theme'
 import { IntlProvider } from 'react-intl'
 
 const properties = require('properties')
