+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/theme/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './theme-container'
+export { default } from './theme'
