+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/theme/theme.tsx
@@ -18,10 +18,10 @@ import {
   ThemeInterface,
   SpecificSizingInterface,
   ThemeColorInterface,
-} from '../../styles/styled-components'
-const user = require('../../../component/singletons/user-instance.js')
-const Common = require('../../../js/Common.js')
-import withListenTo, { WithBackboneProps } from '../../backbone-container'
+} from '../styles/styled-components'
+const user = require('../../component/singletons/user-instance.js')
+const Common = require('../../js/Common.js')
+import withListenTo, { WithBackboneProps } from '../backbone-container'
 const $ = require('jquery')
 const _ = require('underscore')
